consumer/requirements.txt
confluent-kafka==2.5.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.35
pandas==2.2.3
pyarrow==17.0.0
supabase==2.7.4
upstash-redis==1.1.0
python-dotenv==1.0.1
pydantic-settings==2.4.0
